ADD SGP40

+++ Sensor/Sensor.h
@@ -1,13 +1,15 @@
 
+#include <Wire.h>
+
 struct Measurements {
   int co2;
   float temperature;
   float humidity;
 };
 
-int sensorSetup(int offset = 0);
+int sensorSetup(float offset = 0);
 int sensorGet(Measurements& measurements);
 
 //SCD41
-int sensorSingleStart(int offset = 0);
+int sensorSingleStart(float offset = 0);
 int sensorSingleGet(Measurements& measurements);
